Match anywhere in the stack.
Using an iterator as we only have IEnumerable available, and creating a
list of pairs requires maintaining state across iterations. I can't see an
easy way to do it with LINQ, so falling back to this implementation for
now so it can be extracted into a new LINQ-style extension method later.